+++ crates/client/src/client.rs
@@ -21,6 +21,7 @@ pub use crate::{
 };
 use anyhow::anyhow;
 pub use entropy_protocol::{sign_and_encrypt::EncryptedSignedMessage, KeyParams};
+use parity_scale_codec::Decode;
 use rand::Rng;
 use std::str::FromStr;
 pub use synedrion::KeyShare;
@@ -39,7 +40,9 @@ use crate::{
     },
     client::entropy::staking_extension::events::{EndpointChanged, ThresholdAccountChanged},
     substrate::{get_registered_details, submit_transaction_with_pair},
-    user::{get_all_signers_from_chain, get_validators_not_signer_for_relay, UserSignatureRequest},
+    user::{
+        self, get_all_signers_from_chain, get_validators_not_signer_for_relay, UserSignatureRequest,
+    },
     Hasher,
 };
 
@@ -426,6 +429,9 @@ async fn jumpstart_inner(
 /// some point in the near future.
 ///
 /// The returned `nonce` must be used when generating a `quote` for the chain.
+///
+/// This wraps [user::request_attestation] to convert the error to a [ClientError] consistant with
+/// other functions in this module
 #[tracing::instrument(
     skip_all,
     fields(
@@ -435,19 +441,25 @@ async fn jumpstart_inner(
 pub async fn request_attestation(
     api: &OnlineClient<EntropyConfig>,
     rpc: &LegacyRpcMethods<EntropyConfig>,
-    attestee: sr25519::Pair,
-) -> Result<Vec<u8>, ClientError> {
-    tracing::debug!("{} is requesting an attestation.", attestee.public());
+    attestee: &sr25519::Pair,
+) -> Result<[u8; 32], ClientError> {
+    Ok(user::request_attestation(api, rpc, attestee).await?)
+}
 
-    let request_attestation = entropy::tx().attestation().request_attestation();
-
-    let result =
-        submit_transaction_with_pair(api, rpc, &attestee, &request_attestation, None).await?;
-    let result_event = result
-        .find_first::<entropy::attestation::events::AttestationIssued>()?
-        .ok_or(crate::errors::SubstrateError::NoEvent)?;
-
-    let nonce = result_event.0;
-
-    Ok(nonce)
+/// Get oracle data headings
+/// This is useful for program developers to know what oracle data is available
+pub async fn get_oracle_headings(
+    api: &OnlineClient<EntropyConfig>,
+    _rpc: &LegacyRpcMethods<EntropyConfig>,
+) -> Result<Vec<String>, ClientError> {
+    let storage_address = entropy::storage().oracle().oracle_data_iter();
+    let mut iter = api.storage().at_latest().await?.iter(storage_address).await?;
+    let mut headings = Vec::new();
+    while let Some(Ok(kv)) = iter.next().await {
+        // Key is: storage_address || 128 bit hash || key
+        let mut input = &kv.key_bytes[32 + 16 + 1..];
+        let heading = String::decode(&mut input)?;
+        headings.push(heading);
+    }
+    Ok(headings)
 }