+++ crates/testing-utils/src/lib.rs
@@ -22,9 +22,7 @@ pub mod create_test_keyshares;
 pub mod helpers;
 mod node_proc;
 pub mod substrate_context;
-pub use entropy_tss::helpers::tests::{
-    jump_start_network_with_signer as jump_start_network, spawn_testing_validators, ChainSpecType,
-};
+pub use entropy_tss::helpers::tests::{spawn_testing_validators, ChainSpecType};
 pub use node_proc::TestNodeProcess;
 pub use substrate_context::*;