+++ crates/threshold-signature-server/tests/jumpstart_register_sign.rs
@@ -21,32 +21,23 @@ use entropy_client::{
     client as test_client, Hasher,
 };
 use entropy_kvdb::clean_tests;
-use entropy_protocol::{decode_verifying_key, RecoverableSignature};
 use entropy_testing_utils::{
-    constants::{AUXILARY_DATA_SHOULD_SUCCEED, TEST_PROGRAM_WASM_BYTECODE},
+    constants::{
+        AUXILARY_DATA_SHOULD_SUCCEED, PREIMAGE_SHOULD_SUCCEED, TEST_PROGRAM_WASM_BYTECODE,
+    },
     helpers::spawn_tss_nodes_and_start_chain,
     ChainSpecType,
 };
 use entropy_tss::helpers::tests::{do_jump_start, initialize_test_logger};
-use ethers_core::{
-    abi::ethabi::ethereum_types::{H160, H256},
-    types::{RecoveryMessage, Transaction, TransactionRequest, U256},
-    utils::{
-        public_key_to_address,
-        rlp::{Decodable, Rlp},
-    },
-};
 use serial_test::serial;
 use sp_core::Pair;
 use sp_keyring::AccountKeyring;
 use subxt::utils::AccountId32;
 use synedrion::k256::ecdsa::VerifyingKey;
 
-const GOERLI_CHAIN_ID: u64 = 5;
-
 #[tokio::test]
 #[serial]
-async fn integration_test_sign_eth_tx() {
+async fn integration_test_register_sign() {
     initialize_test_logger().await;
     clean_tests();
 
@@ -56,6 +47,7 @@ async fn integration_test_sign_eth_tx() {
     // First jumpstart the network
     do_jump_start(&api, &rpc, AccountKeyring::Alice.pair()).await;
 
+    // Now register an account
     let account_owner = AccountKeyring::Ferdie.pair();
     let signature_request_author = AccountKeyring::One;
 
@@ -84,68 +76,28 @@ async fn integration_test_sign_eth_tx() {
     .await
     .unwrap();
 
-    let eth_verifying_key = decode_verifying_key(&verifying_key).unwrap();
-    let transaction_request = create_unsigned_eth_tx(eth_verifying_key);
-
-    let message = transaction_request.rlp_unsigned().to_vec();
-    let message_hash = Hasher::keccak(&message);
-
+    // Sign a message
     let recoverable_signature = test_client::sign(
         &api,
         &rpc,
         signature_request_author.pair(),
         verifying_key,
-        message,
+        PREIMAGE_SHOULD_SUCCEED.to_vec(),
         Some(AUXILARY_DATA_SHOULD_SUCCEED.to_vec()),
     )
     .await
     .unwrap();
 
+    // Check the signature
+    let message_should_succeed_hash = Hasher::keccak(PREIMAGE_SHOULD_SUCCEED);
     let recovery_key_from_sig = VerifyingKey::recover_from_prehash(
-        &message_hash,
+        &message_should_succeed_hash,
         &recoverable_signature.signature,
         recoverable_signature.recovery_id,
     )
     .unwrap();
-    assert_eq!(eth_verifying_key, recovery_key_from_sig);
-
-    let ethers_signature = recoverable_signature_to_ethers_signature(recoverable_signature);
-
-    // Check the signature
-    let recovered_eth_address =
-        ethers_signature.recover(RecoveryMessage::Hash(H256(message_hash))).unwrap();
-    assert_eq!(recovered_eth_address, public_key_to_address(&eth_verifying_key));
-
-    let signed_transaction_bytes = transaction_request.rlp_signed(&ethers_signature);
-    let rlp = Rlp::new(&signed_transaction_bytes);
-    let transaction = Transaction::decode(&rlp).unwrap();
-
-    // To be sure that the message we are verifying, matches the message that we signed, convert
-    // Transaction back into transaction request
-    let back_into_transaction_request: TransactionRequest = (&transaction).into();
-    // Check that the hashes match
-    assert_eq!(message_hash, Hasher::keccak(&back_into_transaction_request.rlp()));
-
-    // Verify the signed Transaction
-    let recovered_eth_address = transaction.recover_from().unwrap();
-    assert_eq!(recovered_eth_address, public_key_to_address(&eth_verifying_key));
-}
-
-/// Convert a k256 Signature and RecoveryId to an ethers Signature
-fn recoverable_signature_to_ethers_signature(
-    recoverable_signature: RecoverableSignature,
-) -> ethers_core::types::Signature {
-    let recovery_id_u64: u64 = recoverable_signature.recovery_id.to_byte().into();
-    let v: u64 = 27 + recovery_id_u64;
-    let r = U256::from_big_endian(&recoverable_signature.signature.r().to_bytes());
-    let s = U256::from_big_endian(&recoverable_signature.signature.s().to_bytes());
-
-    ethers_core::types::Signature { r, s, v }
-}
-
-/// Create a mock Ethereum transaction request
-fn create_unsigned_eth_tx(verifying_key: VerifyingKey) -> TransactionRequest {
-    let from = public_key_to_address(&verifying_key);
-    let to = H160::zero();
-    TransactionRequest::pay(to, 1000).from(from).chain_id(GOERLI_CHAIN_ID)
+    assert_eq!(
+        verifying_key.to_vec(),
+        recovery_key_from_sig.to_encoded_point(true).to_bytes().to_vec()
+    );
 }