+++ pallets/staking/src/mock.rs
@@ -16,6 +16,7 @@
 use core::convert::{TryFrom, TryInto};
 use std::cell::RefCell;
 
+use entropy_shared::QuoteContext;
 use frame_election_provider_support::{
     bounds::{ElectionBounds, ElectionBoundsBuilder},
     onchain, SequentialPhragmen, VoteWeight,
@@ -400,6 +401,7 @@ impl entropy_shared::AttestationHandler<AccountId> for MockAttestationHandler {
         _x25519_public_key: entropy_shared::X25519PublicKey,
         _provisioning_certification_key: entropy_shared::BoundedVecEncodedVerifyingKey,
         quote: Vec<u8>,
+        _context: QuoteContext,
     ) -> Result<(), sp_runtime::DispatchError> {
         let quote: Result<[u8; 32], _> = quote.try_into();
         match quote {
@@ -440,6 +442,7 @@ pub fn new_test_ext() -> sp_io::TestExternalities {
         ],
         proactive_refresh_data: (vec![], vec![]),
         mock_signer_rotate: (false, vec![], vec![]),
+        jump_started_signers: None,
     };
     pallet_balances.assimilate_storage(&mut t).unwrap();
     pallet_staking_extension.assimilate_storage(&mut t).unwrap();